restructure, testing use of ankiport_core
