+++ server.py
@@ -1,5 +1,6 @@
 from flask import Flask
 import json
+import ankiport_core.quizlet_helper as qh
 
 app = Flask(__name__)
 
@@ -12,6 +13,11 @@ def hello():
 @app.route("/yeet")
 def yeet():
     return json.dumps({"good_word": "to HELL with georgia!"})
+
+
+@app.route("/test")
+def test():
+    return qh.apiTest("hello ")
 
 
 if __name__ == '__main__':
